Use a threshold for product changelog reindexing.
When there's a small number of products to reindex, the tree index may not
be worthwhile - it takes only a portion of a second to generate, but
doesn't save that much with a limited number of rows.

This doesn't apply to categories; assuming a large enough product catalog,
reindexing even one category will benefit.